SupportType query made public
