+++ Facade_Engine/Compute/DeflectionCheck.cs
@@ -20,23 +20,12 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Geometry;
-using BH.oM.Dimensional;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Analytical.Elements;
-using BH.oM.Facade.Elements;
-using BH.oM.Facade.SectionProperties;
 using BH.Engine.Geometry;
 using BH.Engine.Spatial;
-using BH.oM.Base;
-using BH.oM.Base.Attributes;
-using System.ComponentModel;
 using BH.oM.Facade.Enums;
 using BH.oM.Structure.Elements;
-using BH.oM.Structure.SectionProperties;
 using BH.oM.Structure.MaterialFragments;
+using BH.oM.Structure.SectionProperties;
 
 namespace BH.Engine.Facade
 {
@@ -94,6 +83,8 @@ namespace BH.Engine.Facade
 
 
         /***************************************************/
+        /****          Private Methods                  ****/
+        /***************************************************/
 
         private static double MomentOfInertia(this Bar bar)
         {
@@ -104,6 +95,8 @@ namespace BH.Engine.Facade
                 return property.Iy;
         }
 
+        /***************************************************/
+
         private static double YoungsModulus(this Bar bar)
         {
             IIsotropic material = bar?.SectionProperty?.Material as IIsotropic;
@@ -113,106 +106,7 @@ namespace BH.Engine.Facade
                 return material.YoungsModulus;
         }
 
-        private static SupportType SupportType(this Bar bar)
-        {
-            bool isStartFix = IsFix(bar?.Release?.StartRelease);
-            bool isStartPin = IsPin(bar?.Release?.StartRelease);
-            bool isStartSlide = IsSlide(bar?.Release?.StartRelease);
-            bool isEndFix = IsFix(bar?.Release?.EndRelease);
-            bool isEndPin = IsPin(bar?.Release?.EndRelease);
-            bool isEndSlide = IsSlide(bar?.Release?.EndRelease);
-
-            if (isStartFix && isEndFix)
-                return oM.Facade.Enums.SupportType.FixFix;
-            else if (isStartPin && isEndPin)
-                return oM.Facade.Enums.SupportType.PinPin;
-            else if ((isStartSlide && isEndPin) || (isStartPin && isEndSlide))
-                return oM.Facade.Enums.SupportType.PinSlide;
-            else
-                return oM.Facade.Enums.SupportType.Undefined;
-        }
-
-        private static bool IsFix(BH.oM.Structure.Constraints.Constraint6DOF support)
-        {
-            if (support == null)
-            {
-                // Fill in
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            if (!IsFullStiffness(support))
-            {
-                // Fill in - weird stiffness error
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            return support.TranslationX == oM.Structure.Constraints.DOFType.Fixed
-                && support.TranslationY == oM.Structure.Constraints.DOFType.Fixed
-                && support.TranslationZ == oM.Structure.Constraints.DOFType.Fixed
-                && support.RotationX == oM.Structure.Constraints.DOFType.Fixed
-                && support.RotationY == oM.Structure.Constraints.DOFType.Fixed
-                && support.RotationZ == oM.Structure.Constraints.DOFType.Fixed;
-        }
-
-        private static bool IsPin(BH.oM.Structure.Constraints.Constraint6DOF support)
-        {
-            if (support == null)
-            {
-                // Fill in
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            if (!IsFullStiffness(support))
-            {
-                // Fill in - weird stiffness error
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            // Rotation around own self (x) not relevant for this exercise
-            return support.TranslationX == oM.Structure.Constraints.DOFType.Fixed
-                && support.TranslationY == oM.Structure.Constraints.DOFType.Fixed
-                && support.TranslationZ == oM.Structure.Constraints.DOFType.Fixed
-                && support.RotationY == oM.Structure.Constraints.DOFType.Free
-                && support.RotationZ == oM.Structure.Constraints.DOFType.Free;
-        }
-
-        private static bool IsSlide(BH.oM.Structure.Constraints.Constraint6DOF support)
-        {
-            if (support == null)
-            {
-                // Fill in
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            if (!IsFullStiffness(support))
-            {
-                // Fill in - weird stiffness error
-                BH.Engine.Base.Compute.RecordError("");
-                return false;
-            }
-
-            // Rotation around own self (x) not relevant for this exercise
-            return support.TranslationX == oM.Structure.Constraints.DOFType.Free
-                && support.TranslationY == oM.Structure.Constraints.DOFType.Fixed
-                && support.TranslationZ == oM.Structure.Constraints.DOFType.Fixed
-                && support.RotationY == oM.Structure.Constraints.DOFType.Free
-                && support.RotationZ == oM.Structure.Constraints.DOFType.Free;
-        }
-
-        private static bool IsFullStiffness(BH.oM.Structure.Constraints.Constraint6DOF support)
-        {
-            return support.RotationalStiffnessX == 0
-                && support.RotationalStiffnessY == 0
-                && support.RotationalStiffnessZ == 0
-                && support.TranslationalStiffnessX == 0
-                && support.TranslationalStiffnessY == 0
-                && support.TranslationalStiffnessZ == 0;
-        }
+        /***************************************************/
     }
 }
